DBLiver-master/requirements.txt
asgiref==3.2.7
boto3==1.14.4
botocore==1.17.4
certifi==2020.4.5.1
Django==3.0.3
django-filter==2.3.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
joblib==0.15.1
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn-pandas==1.8.0
sqlparse==0.3.1
threadpoolctl==2.1.0
urllib3==1.25.9
whitenoise==5.1.0
wincertstore==0.2
